Added sync configuration to allow login
