+++ opensrp-chw-core/src/main/res/values/strings.xml
@@ -1,5 +1,17 @@
 <resources>
+    <!--Login -->
     <string name="powered_by">Powered by</string>
     <string name="username">Username</string>
     <string name="password">Password</string>
+
+    <!-- Date manipulation-->
+    <string name="day">day</string>
+    <string name="days">days</string>
+    <string name="week">week</string>
+    <string name="weeks">weeks</string>
+    <string name="month">month</string>
+    <string name="months">months</string>
+    <string name="year">year</string>
+    <string name="years">years</string>
+
 </resources>
