+++ opensrp-chw-hf/build.gradle
@@ -19,6 +19,12 @@ android {
         buildConfigField "boolean", 'SUPPORT_QR', 'true'
         buildConfigField "int", "DATABASE_VERSION", '10'
         buildConfigField "int", "DATA_SYNC_DURATION_MINUTES", '15'
+        buildConfigField "long", "MAX_SERVER_TIME_DIFFERENCE", "1800000l"
+        buildConfigField "boolean", "TIME_CHECK", "false"
+        buildConfigField "String", "SYNC_TYPE", '"teamId"'
+        buildConfigField "int", "MAX_SYNC_RETRIES", '3'
+        buildConfigField "long", "BUILD_TIMESTAMP", System.currentTimeMillis() + "L"
+        buildConfigField "boolean", "IS_SYNC_SETTINGS", "false"
     }
 
     buildTypes {
