+++ opensrp-chw-hf/src/main/java/com/opensrp/chw/hf/activity/MainActivity.java
@@ -1,7 +1,7 @@
 package com.opensrp.chw.hf.activity;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
 
 import com.opensrp.hf.R;
