+++ opensrp-chw/src/main/java/org/smartregister/chw/provider/ChwDueRegisterProvider.java
@@ -26,6 +26,7 @@ import org.smartregister.view.contract.SmartRegisterClient;
 
 import java.util.Set;
 
+import static com.opensrp.chw.core.utils.Utils.actualDaysBetweenDateAndNow;
 import static org.smartregister.family.util.Utils.getName;
 
 public class ChwDueRegisterProvider extends FamilyDueRegisterProvider {
@@ -74,7 +75,7 @@ public class ChwDueRegisterProvider extends FamilyDueRegisterProvider {
         String lastVisit = Utils.getValue(pc.getColumnmaps(), ChildDBConstants.KEY.LAST_HOME_VISIT, false);
         if (StringUtils.isNotBlank(lastVisit)) {
             // String lastVisitString = Utils.actualDuration(context, Utils.getDuration(lastVisit));
-            String lastVisitString = Utils.actualDaysBetweenDateAndNow(context, lastVisit);
+            String lastVisitString = actualDaysBetweenDateAndNow(context, lastVisit);
             viewHolder.lastVisit.setText(String.format(context.getString(R.string.last_visit_prefix), lastVisitString));
             viewHolder.lastVisit.setVisibility(View.VISIBLE);
         } else {
