+++ opensrp-chw/src/main/res/values/strings.xml
@@ -124,15 +124,6 @@
     <string name="home_visit">Home Visit</string>
     <string name="profile_activity_not_visited">Marked not visited on %s</string>
 
-    <string name="day">day</string>
-    <string name="days">days</string>
-    <string name="week">week</string>
-    <string name="weeks">weeks</string>
-    <string name="month">month</string>
-    <string name="months">months</string>
-    <string name="year">year</string>
-    <string name="years">years</string>
-
     <string name="family">%s Family</string>
     <string name="remove_entire_family">Removing entire family</string>
     <string name="remove_warning_family">Are you sure you want to remove the %s Family\'s record? This will remove the family\'s entire record AND all of the %s family members\' health records from your device. This action cannot be undone.</string>
